requirements.txt
langchain>=0.1.0
langchain-community>=0.0.10
langchain-core>=0.1.0
langchain-qdrant>=0.0.1
langchain-google-genai>=0.0.5
qdrant-client>=1.7.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
PyPDF2>=3.0.0
streamlit>=1.32.0
pypdf>=3.17.0
watchdog>=3.0.0
